+++ application/routes/file_routes.go
@@ -1,14 +1,14 @@
 package routes
 
 import (
-	"github.com/Uallessonivo/go_card_manager/application/handler"
-	"github.com/Uallessonivo/go_card_manager/domain/interfaces"
+	"github.com/Uallessonivo/go_card_manager/internal/adapters/handlers"
+	"github.com/Uallessonivo/go_card_manager/internal/core/ports"
 	"github.com/gofiber/fiber/v2"
 )
 
-func FileRoutes(app *fiber.App, us interfaces.FileUseCaseInterface) {
-	fileHandler := &handler.FileHandler{
-		UseCase: us,
+func FileRoutes(app *fiber.App, us ports.FileService) {
+	fileHandler := &handlers.FileHandler{
+		FileService: us,
 	}
 
 	app.Post("/file/upload", fileHandler.UploadFile)
