+++ application/routes/user_routes.go
@@ -1,14 +1,14 @@
 package routes
 
 import (
-	"github.com/Uallessonivo/go_card_manager/application/handler"
-	"github.com/Uallessonivo/go_card_manager/domain/interfaces"
+	"github.com/Uallessonivo/go_card_manager/internal/adapters/handlers"
+	"github.com/Uallessonivo/go_card_manager/internal/core/ports"
 	"github.com/gofiber/fiber/v2"
 )
 
-func UserRoutes(app *fiber.App, us interfaces.UserUseCaseInterface) {
-	httpHandler := &handler.UserHandler{
-		UseCase: us,
+func UserRoutes(app *fiber.App, us ports.UserService) {
+	httpHandler := &handlers.UserHandler{
+		UserService: us,
 	}
 	app.Post("/user/create", httpHandler.CreateUser)
 	app.Get("/user/find/id/:id", httpHandler.GetUserByID)
