+++ application/utils/employee_response.go
@@ -1,20 +1,20 @@
 package utils
 
 import (
-	"github.com/Uallessonivo/go_card_manager/domain/entities"
+	entities2 "github.com/Uallessonivo/go_card_manager/internal/core/domain/models"
 )
 
-func EmployeeResponse(dataEmployees []*entities.Employee, dataCards []*entities.Card) []*entities.EmployeeResponse {
-	employeeCards := make(map[string][]*entities.Card)
+func EmployeeResponse(dataEmployees []*entities2.Employee, dataCards []*entities2.Card) []*entities2.EmployeeResponse {
+	employeeCards := make(map[string][]*entities2.Card)
 
 	for _, card := range dataCards {
 		employeeCards[card.Owner] = append(employeeCards[card.Owner], card)
 	}
 
-	var employeeResponses []*entities.EmployeeResponse
+	var employeeResponses []*entities2.EmployeeResponse
 
 	for _, employee := range dataEmployees {
-		employeeResponse := &entities.EmployeeResponse{
+		employeeResponse := &entities2.EmployeeResponse{
 			ID:   employee.ID,
 			Name: employee.Name,
 			Cpf:  employee.Cpf,
