--- domain/entities/auth.go
@@ -1,10 +1,0 @@
-package entities
-
-type LoginRequest struct {
-	Email    string `json:"email"`
-	Password string `json:"password"`
-}
-
-type LoginResponse struct {
-	Token string `json:"token"`
-}
