--- domain/entities/file.go
@@ -1,6 +1,0 @@
-package entities
-
-type UploadResponse struct {
-	Message     string
-	FailedCards []*CardRequest
-}
