--- domain/interfaces/auth_interface.go
@@ -1,9 +1,0 @@
-package interfaces
-
-import "github.com/Uallessonivo/go_card_manager/domain/entities"
-
-type AuthUseCaseInterface interface {
-	Login(input *entities.LoginRequest) (*entities.LoginResponse, error)
-	GenerateJWT() error
-	ValidateJWT() error
-}
