+++ internal/core/domain/errors/errors.go
