+++ internal/core/domain/models/auth.go
@@ -1,0 +1,10 @@
+package models
+
+type LoginRequest struct {
+	Email    string `json:"email"`
+	Password string `json:"password"`
+}
+
+type LoginResponse struct {
+	Token string `json:"token"`
+}
