+++ internal/core/domain/models/file.go
@@ -1,0 +1,6 @@
+package models
+
+type UploadResponse struct {
+	Message     string
+	FailedCards []*CardRequest
+}
